Fix React warnings in console for T&C modal
Due to the duplication of the T&C content, the were some undetected
warning message in the console, even thought it seemed to work. With
small adjustments to the JSX template those could be silenced. The
remaining warning do not appear to be caused by us but dependencies.

Related to: #34