Update documentation and restructure it
There are no actual user specific information in the README. Therefore
the technical instructions have been moved to the developer
documentation introduced by #48. The main README that gets initially
displayed when landing on the project gives just a short introduction to
the repository and links to the README-dev.

Related to: trustlines-network/project#798