Deploy the website from CircleCI
This uses the `deploy-www.trustlines.foundation` context to configure
a custom ssh key that needs to have write permissions, in order to run the
`yarn deploy` command for commits from the master branch.

I've been thinking about using a dedicated release branch. This would
give us a bit more safety, but would also mean a bit more work for
merging to that branch.

So, let's try to use this without a dedicated release branch and see
how that works out.