Implement basic auction bid flow screens
The auction web-app should support the user to participate. The
implemented screens describe the basic flow of the application. The
transition between the screen works by button clicks, partially
triggering a loading screen for 2 seconds. There is no further logic
integrated so far.

The new added web-app can be disabled by the configuration. This is
enabled per default, but disabled for the productive stage.

Related to: #69