Implement basic bid flow
Implement the basic bid flow. This allowes bidding in the ideal case,
when the correct account is selected. It will already warn if on the
wrong chain or no web3 can be found. It will also already only ask for
permission if it was not given yet, or only ask for allowance if there
is not yet. However, a lot of different errors can not be handled yet,
including not whitelisted, not enough TLN, switching accounts in the
middle of the process, remove permission in the middle of the process.