Better handle not rdy to bid state in auction flow
Before the state if a certain account can bid was only checked at
certain points. This could lead to problems when the address was changed
in the web3 browser during the flow. We now currently check the status
for things like allowance, whitelisted, already bid and change the view
accordingly.